// src/lib/api/v1.d.ts
export interface components {
    schemas: {
// ...
        TransitRoute: {
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            agencyId: string;
            /**
             * @description Kerékpár szállítása engedélyezett-e a járaton.
             * @example false
             */
            bikesAllowed: boolean;
            /**
             * @deprecated
             * @description A járat színe. Deprecated: használjuk a `style` attribútumot helyette.
             * @example FFD800
             */
            color: string;
            /**
             * @description A járat neve. Ha egy | szerepel a névben, akkor a végállomásokat választja el amelyek külön sorokban megjeleníthetőek. COMPACT referenciák kérése esetében nincs kitöltve.
             * @example Újbuda-központ M | Széll Kálmán tér M
             */
            description?: string;
            /**
             * @deprecated
             * @description A járat ikonjának megjelenítési szövege. Deprecated: használjuk a `style` attribútumot helyette.
             * @example 4
             */
            iconDisplayText?: string;
            /**
             * @deprecated
             * @description A járat ikonjának megjelenítési típusa. Deprecated: használjuk a `style` attribútumot helyette.
             * @example BOX
             */
            iconDisplayType?: string;
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            id: string;
            /** @description A járat hosszú neve. */
            longName?: string;
            /**
             * @description A járat rövid neve.
             * @example 4
             */
            shortName?: string;
            /**
             * Format: int32
             * @description A járat rendezési sorszáma a többi járathoz viszonyítva.
             * @example 10
             */
            sortOrder: number;
            style: components["schemas"]["TransitRouteStyle"];
            /**
             * @deprecated
             * @description A járat szövegének színe. Deprecated: használjuk a `style` attribútumot helyette.
             * @example 000000
             */
            textColor: string;
            type: components["schemas"]["TraverseMode"];
            /** @description A járat linkje. */
            url?: string;
        };
// ...
        TransitTrip: {
            /**
             * @description Szállítható-e kerékpár a meneten.
             * @example false
             */
            bikesAllowed: boolean;
            /**
             * @description A menethez tartozó GTFS block_id, amennyiben létezik.
             * @example C31350_3040_5_15
             */
            blockId?: string;
            /**
             * @description A menet irányának azonosítója.
             * @example 1
             */
            directionId?: string;
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            id: string;
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            routeId: string;
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            serviceId: string;
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            shapeId: string;
            /**
             * @description A menet célállomása.
             * @example Széll Kálmán tér M
             */
            tripHeadsign: string;
            /**
             * @description A menet rövid neve.
             * @example 4
             */
            tripShortName?: string;
            /**
             * @description Alacsonypadlós-e a menet.
             * @example true
             */
            wheelchairAccessible: boolean;
        };
// ...
        TransitVehicle: {
            /**
             * Format: float
             * @description A jármű irányszöge.
             * @example 97
             */
            bearing?: number;
            capacity?: components["schemas"]["TransitVehicleOccupancy"];
            congestionLevel: components["schemas"]["CongestionLevel"];
            /**
             * @description Igaz, ha a jármű letért az útvonaláról.
             * @example false
             */
            deviated: boolean;
            /**
             * @description A jármű kijelzőjén megjelenő célállomáskép.
             * @example Széll Kálmán tér M
             */
            label?: string;
            /**
             * Format: int64
             * @description A járműhöz tartozó utolsó valós idejű adat időbélyege másodpercben.
             * @example 1625683727
             */
            lastUpdateTime: number;
            /**
             * @description A jármű rendszáma.
             * @example V2035
             */
            licensePlate?: string;
            location: components["schemas"]["TransitCoordinatePoint"];
            /**
             * @description A jármű típusa.
             * @example Siemens Combino
             */
            model?: string;
            occupancy?: components["schemas"]["TransitVehicleOccupancy"];
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            routeId?: string;
            /**
             * @description Menetrendi nap
             * @example 20241231
             */
            serviceDate?: string;
            /**
             * @description Igaz, ha a jármű pozíciója kiöregedett.
             * @example false
             */
            stale?: boolean;
            status?: components["schemas"]["VehicleStatus"];
            /**
             * Format: int32
             * @description Hol tart a jármű a két megálló közti szakaszon, százalékban.
             * @example 28
             */
            stopDistancePercent?: number;
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            stopId?: string;
            /**
             * Format: int32
             * @description A jármű által érintett aktuális megálló sorszáma a meneten.
             * @example 1
             */
            stopSequence?: number;
            style?: components["schemas"]["TransitVehicleStyle"];
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            tripId?: string;
            /**
             * @description Entitás azonosító
             * @example BKK_DF544
             */
            vehicleId: string;
            /**
             * @deprecated
             * @description A jármű fajtája. Deprecated: használjuk a `style` attribútumot.
             * @example TRAM
             * @enum {string}
             */
            vehicleRouteType?: "BICYCLE" | "CAR" | "WALK" | "BUS" | "CABLE_CAR" | "COACH" | "FERRY" | "FUNICULAR" | "GONDOLA" | "RAIL" | "SUBURBAN_RAILWAY" | "SUBWAY" | "TRAM" | "TROLLEYBUS" | "LOCAL" | "TRANSIT";
            /**
             * @description A menet utazástervező azonosítója, amelyet a `fromPlace` megadásához lehet használni.
             * @example 20210708:BKK:AS4356
             */
            vertex?: string;
            /**
             * @description Igaz, ha a jármű alacsonypadlós.
             * @example true
             */
            wheelchairAccessible: boolean;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}